Move oauth entity plugin config
